add isoptional support to client api
