+++ src/Routine/Core/DataModel.cs
@@ -15,6 +15,8 @@ namespace Routine.Core
         public DataModel() { }
         public DataModel(IDictionary<string, object> model)
         {
+            if(model == null) return;
+
             if (model.TryGetValue("Marks", out var marks))
             {
                 Marks = ((IEnumerable)marks).Cast<string>().ToList();
