+++ src/Routine/Core/InitializerModel.cs
@@ -14,6 +14,8 @@ namespace Routine.Core
         public InitializerModel() { }
         public InitializerModel(IDictionary<string, object> model)
         {
+            if(model == null) return;
+
             if (model.TryGetValue("Marks", out var marks))
             {
                 Marks = ((IEnumerable)marks).Cast<string>().ToList();
