+++ src/Routine/Core/ObjectModel.cs
@@ -24,6 +24,8 @@ namespace Routine.Core
         public ObjectModel() { }
         public ObjectModel(IDictionary<string, object> model)
         {
+            if(model == null) return;
+
             if (model.TryGetValue("Id", out var id))
             {
                 Id = (string)id;
