+++ src/Routine/Core/OperationModel.cs
@@ -16,6 +16,8 @@ namespace Routine.Core
         public OperationModel() { }
         public OperationModel(IDictionary<string, object> model)
         {
+            if(model == null) return;
+
             if (model.TryGetValue("Marks", out var marks))
             {
                 Marks = ((IEnumerable)marks).Cast<string>().ToList();
