+++ src/Routine/Core/VariableData.cs
@@ -12,6 +12,8 @@ namespace Routine.Core
         public VariableData() { }
         public VariableData(IDictionary<string, object> data)
         {
+            if(data == null) return;
+
             if (data.TryGetValue("IsList", out var isList))
             {
                 IsList = (bool)isList;
