+++ test/Routine.Test/Core/CoreTestBase.cs
@@ -194,13 +194,35 @@ namespace Routine.Test.Core
         protected ParameterModel PModel(string name, string viewModelId, params int[] groups) => PModel(name, viewModelId, false, groups);
         protected ParameterModel PModel(string name, bool isList, params int[] groups) => PModel(name, DefaultObjectModelId, isList, groups);
         protected ParameterModel PModel(string name, string viewModelId, bool isList, params int[] groups) =>
-            new()
-            {
-                Name = name,
-                IsList = isList,
-                ViewModelId = viewModelId,
-                Groups = groups.Any() ? groups.ToList() : new List<int> { 0 }
-            };
+            PModel(name, viewModelId: viewModelId, isList: isList, defaultValue: null, groups: groups);
+
+        protected ParameterModel PModel(string name,
+            string viewModelId = null,
+            bool isList = false,
+            string defaultValue = null,
+            int[] groups = null
+        ) => new()
+        {
+            Name = name,
+            IsList = isList,
+            ViewModelId = viewModelId ?? DefaultObjectModelId,
+            Groups = (groups ?? new[] { 0 }).ToList(),
+            DefaultValue = defaultValue == null
+                ? null
+                : new()
+                {
+                    IsList = isList,
+                    Values = new()
+                    {
+                        new()
+                        {
+                            Id = defaultValue,
+                            ModelId = viewModelId ?? DefaultObjectModelId
+                        }
+                    }
+                },
+            IsOptional = defaultValue != null
+        };
 
         #endregion
